furthe qa changes

+++ routes/roles.js
@@ -85,8 +85,14 @@ router.get("/", passport.authenticate("jwt", { session: false }), (req, res, nex
  *     responses:
  *       200:
  *         description: The newly created role
+ *       400:
+ *         description: Bad Request
+ *       401:
+ *         description: Unauthorized
  *       403:
  *        description: Forbidden due to capability requirements
+ *       500:
+ *         description: Internal Server Error
  */
 router.post(
     "/create",
@@ -97,6 +103,10 @@ router.post(
         MiddlewareHelper.userHasCapability("Hall Monitor"),
     ],
     (req, res, next) => {
+        if (!req.body.name || !req.body.description || !req.body.authoriser || !req.body.capabilities) {
+            res.status(400).json({ success: false, msg: "Incorrect Parameters" });
+            return;
+        }
         const payload = req.body;
         RoleModel.create(
             {
@@ -186,9 +196,15 @@ router.post(
  *      - application/json
  *     responses:
  *       200:
- *         description: The updated role
+ *         description: The newly created role
+ *       400:
+ *         description: Bad Request
+ *       401:
+ *         description: Unauthorized
  *       403:
  *        description: Forbidden due to capability requirements
+ *       500:
+ *         description: Internal Server Error
  */
 router.put(
     "/update",
@@ -199,6 +215,10 @@ router.put(
         MiddlewareHelper.userHasCapability("Hall Monitor"),
     ],
     (req, res, next) => {
+        if (!req.body.id || !req.body.name || !req.body.description || !req.body.authoriser || !req.body.capabilities) {
+            res.status(400).json({ success: false, msg: "Incorrect Parameters" });
+            return;
+        }
         // Get all capabilities
         const payload = req.body;
         RoleModel.updateByPrimaryKey(
@@ -283,8 +303,14 @@ router.put(
  *     responses:
  *       200:
  *         description: Success status
+ *       400:
+ *         description: Bad Request
+ *       401:
+ *         description: Unauthorized
  *       403:
- *        description: Forbidden due to capability requirements
+ *         description: Forbidden due to capability requirements
+ *       500:
+ *         description: Internal Server Error
  */
 router.post(
     "/links/sync",
@@ -295,6 +321,10 @@ router.post(
         MiddlewareHelper.userHasCapability("Hall Monitor"),
     ],
     (req, res, next) => {
+        if (!req.body.roles || !req.body.link_id || !req.body.link_type) {
+            res.status(400).json({ success: false, msg: "Incorrect Parameters" });
+            return;
+        }
         // Get params
         const payload = req.body;
 
@@ -315,7 +345,7 @@ router.post(
                     res.status(500).json({ success: false, msg: err });
                     return;
                 }
-                res.json({ success: true, msg: "Role links synced!" });
+                res.json({ success: true, msg: "Role links synced" });
             }
         );
     }
@@ -351,8 +381,14 @@ router.post(
  *     responses:
  *       200:
  *         description: Success status
+ *       400:
+ *         description: Bad Request
+ *       401:
+ *         description: Unauthorized
  *       403:
- *        description: Forbidden due to capability requirements
+ *         description: Forbidden due to capability requirements
+ *       500:
+ *         description: Internal Server Error
  */
 router.post(
     "/links/create",
@@ -363,6 +399,10 @@ router.post(
         MiddlewareHelper.userHasCapability("Hall Monitor"),
     ],
     (req, res, next) => {
+        if (!req.body.role_id || !req.body.link_id || !req.body.link_type) {
+            res.status(400).json({ success: false, msg: "Incorrect Parameters" });
+            return;
+        }
         // Get params
         const payload = req.body;
 
@@ -383,7 +423,7 @@ router.post(
                     res.status(500).json({ success: false, msg: err });
                     return;
                 }
-                res.json({ success: true, msg: "Role link created!" });
+                res.json({ success: true, msg: "Role link created" });
             }
         );
     }
@@ -419,8 +459,14 @@ router.post(
  *     responses:
  *       200:
  *         description: Success status
+ *       400:
+ *         description: Bad Request
+ *       401:
+ *         description: Unauthorized
  *       403:
- *        description: Forbidden due to capability requirements
+ *         description: Forbidden due to capability requirements
+ *       500:
+ *         description: Internal Server Error
  */
 router.delete(
     "/links/delete",
@@ -431,6 +477,10 @@ router.delete(
         MiddlewareHelper.userHasCapability("Hall Monitor"),
     ],
     (req, res, next) => {
+        if (!req.body.role_id || !req.body.link_id || !req.body.link_type) {
+            res.status(400).json({ success: false, msg: "Incorrect Parameters" });
+            return;
+        }
         // Get params
         const payload = req.body;
 
@@ -441,7 +491,7 @@ router.delete(
                 res.status(500).json({ success: false, msg: err });
                 return;
             }
-            res.json({ success: true, msg: "Role link deleted!" });
+            res.json({ success: true, msg: "Role link deleted" });
         });
     }
 );
@@ -466,6 +516,12 @@ router.delete(
  *     responses:
  *       200:
  *         description: The role
+ *       401:
+ *         description: Unauthorized
+ *       404:
+ *         description: Not Found
+ *       500:
+ *         description: Internal Server Error
  */
 router.get(
     "/:id",
@@ -477,7 +533,11 @@ router.get(
             if (err) {
                 res.status(500).json({ success: false, msg: err });
             } else {
-                res.json(result);
+                if (result) {
+                    res.json(result);
+                } else {
+                    res.status(404).json({ success: false, msg: "Role not found" });
+                }
             }
         });
     }
@@ -502,9 +562,15 @@ router.get(
  *      - application/json
  *     responses:
  *       200:
- *         description: Role deletion status
+ *         description: The role
+ *       401:
+ *         description: Unauthorized
+ *       404:
+ *         description: Not Found
  *       403:
  *        description: Forbidden due to capability requirements
+ *       500:
+ *         description: Internal Server Error
  */
 router.delete(
     "/delete",
@@ -515,12 +581,23 @@ router.delete(
         MiddlewareHelper.userHasCapability("Hall Monitor"),
     ],
     (req, res, next) => {
-        // Get all capabilities
-        RoleModel.deleteByPrimaryKey(req.body.id, (err, result) => {
-            if (err) {
-                res.status(500).json({ success: false, msg: err });
+        if (!req.body.id) {
+            res.status(400).json({ success: false, msg: "Incorrect Parameters" });
+            return;
+        }
+        RoleModel.getByPrimaryKey(req.params.id, (errGet, resultGet) => {
+            if (errGet) {
+                res.status(500).json({ success: false, msg: errGet });
+            } else if (resultGet) {
+                RoleModel.deleteByPrimaryKey(req.body.id, (err, result) => {
+                    if (err) {
+                        res.status(500).json({ success: false, msg: err });
+                    } else {
+                        res.json({ success: true, msg: "Role deleted successfully" });
+                    }
+                });
             } else {
-                res.json({ success: true, msg: "Role deleted successfully!" });
+                res.status(404).json({ success: false, msg: "Role not found" });
             }
         });
     }
